webpacker使用に伴うファイル移行と調整
