+++ Gemfile
@@ -59,6 +59,7 @@ group :development do
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
   gem 'spring'
   gem 'spring-watcher-listen', '~> 2.0.0'
+  gem 'foreman'
 end
 
 group :test do
@@ -67,8 +68,5 @@ end
 
 # Windows does not include zoneinfo files, so bundle the tzinfo-data gem
 gem 'tzinfo-data', platforms: [:mingw, :mswin, :x64_mingw, :jruby]
-gem "sass-rails", "~>5.0"
-gem "bootstrap"
-gem "jquery-rails"
 gem "jquery-ui-rails"
 gem 'webpacker', '~> 5.0'
