+++ Procfile
@@ -1,0 +1,2 @@
+rails: rails s -p 3000
+webpacker: ./bin/webpack-dev-server
