+++ app/helpers/application_helper.rb
@@ -1,6 +1,6 @@
 module ApplicationHelper
 	def embedded_svg filename, options={}
-		file = File.read(Rails.root.join('app', 'assets', 'images', filename))
+		file = File.read(Rails.root.join('app', 'webpacker_assets', 'images', filename))
 		doc = Nokogiri::HTML::DocumentFragment.parse file
 		svg = doc.at_css 'svg'
 		if options[:class].present?
