+++ app/views/static_pages/top.html.erb
@@ -103,5 +103,4 @@
 		<%= link_to "Signup", new_user_path, class: 'btn btn-outline-success btn-lg section-signup' %>
 	</div>
 </section>
-
-
+<%= javascript_pack_tag 'src/index' %>
