+++ app/webpacker_assets/images/top-bg.jpg
