+++ app/webpacker_assets/packs/application.js
@@ -12,7 +12,11 @@
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
 // or the `imagePath` JavaScript helper below.
 //
-// const images = require.context('../images', true)
-// const imagePath = (name) => images(name, true)
+const images = require.context('../images', true)
+const imagePath = (name) => images(name, true)
 
-console.log('Hello World from Webpacker')
+import Rails from 'rails-ujs'
+Rails.start()
+import '../stylesheets/application'
+import 'bootstrap'
+import 'jquery'
