+++ app/webpacker_assets/packs/src/index.js
@@ -1,0 +1,1 @@
+import "../../src/static_pages";
