+++ app/webpacker_assets/stylesheets/application.scss
@@ -14,9 +14,12 @@
  *= require_tree .
  */
 
-@import "bootstrap";
 @import "./default.scss";
 @import "./variable";
+@import "./default.scss";
+@import "./static_pages.scss";
+@import "./users";
+@import "~bootstrap/scss/bootstrap";
 
 // ここからナビバー設定
 .navbar-light .navbar-toggler {
@@ -27,7 +30,7 @@
 
 .navbar-light .navbar-toggler-icon {
     background-image: url("data:image/svg+xml,%3csvg viewBox='0 0 30 30'
-    xmlns='http://www.w3.org/2000/svg'%3e%3cpath stroke='rgba(88, 202, 180, 1)' stroke-width='3' stroke-linecap='round' stroke-miterlimit='5' d='M4 7h22M4 15h22M4 23h22'/%3e%3c/svg%3e");
+    // xmlns='http://www.w3.org/2000/svg'%3e%3cpath stroke='rgba(88, 202, 180, 1)' stroke-width='3' stroke-linecap='round' stroke-miterlimit='5' d='M4 7h22M4 15h22M4 23h22'/%3e%3c/svg%3e");
 
 }
