+++ config/webpack/environment.js
@@ -1,3 +1,4 @@
 const { environment } = require('@rails/webpacker')
 
 module.exports = environment
+environment.loaders.get('sass').use.push('import-glob-loader')
